Move highlight logic to a function
